Break subclasses into correct namespace
